Add pre-commit
This commit adds a pre-commit configuration file to the project, which
includes hooks for end-of-file-fixer, trailing-whitespace, and
requirements-txt-fixer from the pre-commit/pre-commit-hooks repository,
as well as black and isort from the psf/black and pycaq/isort
repositories, respectively. The isort hook is configured to use the black
profile.

Additionally, the requirements.txt file is updated to include the
pre-commit package with version >=3.3.3.