Add sphinx-copybutton extension to Sphinx configuration
This commit adds the sphinx-copybutton extension to the list of extensions in the Sphinx configuration file (conf.py). This extension adds a "copy" button to each code block in the Sphinx documentation, making it easier for users to copy code examples. The sphinx-copybutton package is also added to the list of dependencies in setup.cfg.