Refine solar irradiation discussion and comparison with `pvlib python`
This commit refines the discussion on solar irradiation in the paper. The focus is
now on direct beam irradiation, which contributes the most to solar irradiation. The
conditions assumed are a clear sky with no clouds and an atmosphere free of pollutants
The commit also updates the comparison of `pysolorie` with other libraries. Instead of
comparing with `Astropy`, the comparison is now with `pvlib python`, a library for
simulating the performance of photovoltaic energy systems. The discussion on backtracking
and shading effect for a field of one-axis sun-trackers, as discussed in the paper by
Anderson & Mikofski, is also included.